% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.cohorts_coded.r
\docType{data}
\name{cohorts_coded}
\alias{cohorts_coded}
\title{Disease cohorts with coded identifiers}
\format{
cohorts_coded - A list object where elements contain all patient 
IDs associated with a given diagnosis, as included in the dataset
Thistlethwaite2020.
}
\usage{
data(cohorts_coded)
}
\description{
Patient sample identifiers mapped to known clinical
diagnoses.
}
\examples{
data(cohorts_coded)
}
\keyword{datasets}
