% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctdTag.R
\name{ctdTag}
\alias{ctdTag}
\title{Tag CTD features}
\usage{
ctdTag(
  file,
  dir,
  dbname = getDatabaseName("ctdTag"),
  tagScheme = NULL,
  height = 550,
  clickDistanceCriterion = 0.02,
  debug = 0L
)
}
\arguments{
\item{file}{character value specifying the name of a file containing CTD data.
This file must be in a format that is handled by \code{\link[oce:read.oce]{oce::read.oce()}}.  You may
not supply \code{dir} if \code{file} is supplied.}

\item{dir}{character value specifying the name of a directory containing
files that hold CTD data. You may not supply \code{file} if \code{dir} is supplied.}

\item{dbname}{optional character value specifying the name of a sqlite
database used to hold tagging information.  If not provided, a file name
is constructed with \code{\link[=getDatabaseName]{getDatabaseName()}}.
If this file does not exist, it is created, and a table named
\code{tagScheme} is constructed in it, according to the \code{tagScheme} parameter.
However, if the file already exists, then its \code{tagScheme} table is
not altered.}

\item{tagScheme}{optional \link{data.frame} containing items \code{tagCode} (integer) and
\code{tagLabel} (character).  If not given, the labels default to
\code{c("iTop", "iTop?", "iBot", "iBot?", "WS", "WS?", "CF", "CF?")} and
the codes default to 1:8.}

\item{height}{height of the plot in pixels.}

\item{clickDistanceCriterion}{numeric value controlling whether a mouse click
selects, or deselects a point.  If the ratio of the distance from the mouse
location to the nearest point exceeds \code{clickDistanceCriterion} times
the span of the current view, then that nearest point is selected. Otherwise,
any existing selected point is deselected.}

\item{debug}{integer indicating whether to print some informative information
to the R console.  Use 0 for mainly silent processing, 1 for more information, etc.}
}
\value{
\code{\link[=ctdTag]{ctdTag()}} returns either an empty string, if the procedure worked, or a description
of the problem, otherwise.
}
\description{
This function may be used to tag features in CTD data.  It works by
launching an R/Shiny app, providing it with information gleaned
from the function parameters.  It may be used to work with a single data
file (specified with \code{file}) or with all the \code{.cnv} files in a directory
(specified with \code{dir}).  The user interface is meant to be reasonably
self-explanatory.  The information provided via the user's mouse clicks
and keyboard actions is recorded in an SQLite database named (by default)
\verb{~/ctdTag_USERNAME.db}, where \code{USERNAME} is the user's login name on the
computer.  This files \emph{must not} be edited or modified by the user, unless
that user has a high degree of expertise in SQL and in the functioning
of this package.  Since \code{ctdTag()} is still in development, sensible users
will make frequent backups of their \verb{~/ctdTag_USERNAME.db} file.
}
\examples{
\dontrun{
library(ctd)
file <- "~/data/arctic/beaufort/2012/d201211_0047.cnv"
ctdTag(file=file)
dir <- "~/data/arctic/beaufort/2012"
ctdTag(dir=dir)
}


}
\author{
Dan Kelley
}
