% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.combineData.r
\name{data.combineData}
\alias{data.combineData}
\title{Combine datasets from different source files.}
\usage{
data.combineData(ref, data)
}
\arguments{
\item{ref}{- Reference samples normalized, imputed data.}

\item{data}{- Normalized, imputed data. Data matrix with observations as rows, features as columns.}
}
\value{
combined.data - Z-transformed data.
}
\description{
Combine datasets from different source files.
}
