% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.imputeData.r
\name{data.imputeData}
\alias{data.imputeData}
\title{Impute missing values as lowest observed value.}
\usage{
data.imputeData(data, ref)
}
\arguments{
\item{data}{- Normalized, imputed data. Data matrix with observations as rows, features as columns.}

\item{ref}{- Reference samples normalized, imputed data.}
}
\value{
imputed.data - Z-transformed data.
}
\description{
Impute missing values as lowest observed value.
}
