% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.surrogateProfiles.r
\name{data.surrogateProfiles}
\alias{data.surrogateProfiles}
\title{Surrogate profiles}
\usage{
data.surrogateProfiles(data, sd = 1)
}
\arguments{
\item{data}{- Data matrix with observations as rows, features as columns.}

\item{sd}{- The level of variability (standard deviation) around each feature's
mean you want to add in surrogate profiles.}
}
\value{
data_mx - Data matrix with added surrogate profiles.
}
\description{
Fill in a data matrix with low n, high p with surrogate profiles.
}
