% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.zscoreData.r
\name{data.zscoreData}
\alias{data.zscoreData}
\title{Z-transform normalized, imputed metabolomics data}
\usage{
data.zscoreData(data, ref)
}
\arguments{
\item{data}{- Normalized, imputed data. Data matrix with observations as rows, features as columns.}

\item{ref}{- Reference samples normalized, imputed data.}
}
\value{
zscore.data - Z-transformed data.
}
\description{
Z-transform normalized, imputed metabolomics data
}
