% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{getDatabaseName}
\alias{getDatabaseName}
\title{Construct a database name.}
\usage{
getDatabaseName(prefix = "ctdTag", username, suffix)
}
\arguments{
\item{prefix}{character value indicating the prefix for the database
name.}

\item{username}{character value for the user's login name. If not provided,
an attempt is made to construct this using \code{\link[=Sys.info]{Sys.info()}}.}

\item{suffix}{optional character value that may be included in the filename,
if provided.}
}
\value{
\link{getDatabaseName} returns a character value constructed by
stringing \code{prefix}, \code{username} (and possibly \code{suffix}) together, separated
by underscore characters, with \code{.db} being added to the end;
see \sQuote{Examples}.
}
\description{
A database name is constructed to in the form \verb{PREFIX_USERNAME[_SUFFIX].db},
where the upper-case items are inferred from corresponding parameters.
}
\examples{
getDatabaseName() # ctdTag_kelley.db, for the author
getDatabaseName("foo", "bar", "01") # foo_bar_01.db

}
\author{
Dan Kelley
}
