% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{getTableFromDatabase}
\alias{getTableFromDatabase}
\title{Get a table from an SQLite database}
\usage{
getTableFromDatabase(table, dbname)
}
\arguments{
\item{table}{character value naming the table of interest.}

\item{dbname}{character value naming the database file.}
}
\description{
Get a table from an SQLite database
}
\author{
Dan Kelley
}
