% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.diffuseP1Movie.r
\name{graph.diffuseP1Movie}
\alias{graph.diffuseP1Movie}
\title{Make a movie of the diffusion of probability, P1, from a starting node.}
\usage{
graph.diffuseP1Movie(p1, startNode, G, visitedNodes, ig,
  recursion_level = 1, output_dir = getwd())
}
\arguments{
\item{p1}{- The probability being dispersed from the starting node, startNode.}

\item{startNode}{- The first variable drawn in the adaptive permutation node sequence, from which p1 gets dispersed.}

\item{G}{- A list of probabilities, with names of the list being the node names in the background knowledge graph.}

\item{visitedNodes}{- A character vector of node names, storing the history of previous draws in the permutation sequence.}

\item{graphNumber}{- If testing against multiple background knowledge graphs, this is the index associated with the adjacency matrix that codes for G. Default value is 1.}
}
\value{
G - A list of returned probabilities after the diffusion of probability has truncated, with names of the list being the node names in the background knowledge graph.
}
\description{
Recursively diffuse probability from a starting node based on the connectivity of the background knowledge graph, representing the likelihood that a variable will be
        most influenced by a perturbation in the starting node.
}
\keyword{diffusion}
\keyword{event}
\keyword{generative}
\keyword{methods}
\keyword{network}
\keyword{walker}
