% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.Miller2015.r
\docType{data}
\name{Miller2015}
\alias{Miller2015}
\title{Untargeted metabolomic analysis for the clinical screening of inborn errors
of metabolism}
\source{
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4626538/bin/10545_2015_9843_MOESM1_ESM.xls}{Dataset}
}
\usage{
data(Miller2015)
}
\value{
Miller2015 - The data matrix (metabolite features are rows, patient observations are columns) for 186 untargeted metabolomics 
                      patient samples, alongside metabolite annotations.

diagnoses - A data.frame where all patient IDs are mapped to their given biochemical diagnosis.
}
\description{
Global metabolic profiling obtained by untargeted mass spectrometry-based
metabolomic platform for the detection of novel and known inborn errors of
metabolism.  This untargeted approach collected z-score values for ~900 unique
compounds (including ~500 named human analytes) from human plasma.  Data set
contains 190 individual plasma samples (120 confirmed inborn errors of
metabolism).  The outcome describes excellent sensitivity and specificity for
the detection of a wide rage of metabolic disorders and identified novel
biomarkers for some diseases.
}
\examples{
require(CTD)
data(Miller2015)
}
\references{
Miller et al. (2015) J Inherit Metab Dis. 2015; 38: 1029–1039
(\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4626538/}{PubMed})
}
\keyword{datasets}
