% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.blowoutPtSim.r
\name{mle.blowoutSim}
\alias{mle.blowoutSim}
\title{Module that best explains the patient similarity assigned between a set of patients.}
\usage{
mle.blowoutSim(patientSim, data_mx, ptIDs, ig_pruned, kmx)
}
\arguments{
\item{patientSim}{- A similarity matrix, where row and columns are patient identifiers.}

\item{data_mx}{- The matrix that gives the perturbation strength (z-scores) for all variables (columns) for each patient (rows).}

\item{ptIDs}{- The identifier associated with patient 1's sample.}

\item{ig_pruned}{- The igraph object associated with the pruned disease+reference differential interaction network.}

\item{kmx}{- The maximum metabolite set size probed when assessing patient similarity.}
}
\value{
ptsim_blowout - An igraph object showing the module blowout describing the similarity between patients in ptIDs.
}
\description{
Module that best explains the patient similarity assigned between a set of patients.
}
\examples{
require(CTD)
data(Thistlethwaite2019)
data_mx = as.matrix(data_mx)
data_mx = suppressWarnings(apply(data_mx, c(1,2), as.numeric))
data_mx = data_mx[,-c(1,2,3,4,5,6,7,8)]
# Load your background network, ig_pruned and your computed patientSim matrix
kmns.clust = kmeans(patientSim, centers=4)
table(kmns.clust$cluster)
ptIDs = names(kmns.clust$cluster[which(kmns.clust$cluster==1)])
ptsim_blowout = mle.blowoutSim(patientSim, data_mx, ptIDs, ig_pruned, kmx=15)
plot.igraph(ptsim_blowout, layout=layout.circle, edge.width=50*abs(E(ptsim_blowout)$weight))
}
