% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.getPermMovie.r
\name{mle.getPermMovie}
\alias{mle.getPermMovie}
\title{Capture the movement of the adaptive walk of the diffusion probability method.}
\usage{
mle.getPermMovie(subset.nodes, ig, output_filepath, movie = TRUE)
}
\arguments{
\item{subset.nodes}{- The subset of variables, S, in a background graph, G.}

\item{ig}{- The igraph object associated with the background knowledge graph.}

\item{movie}{- If you want to make a movie, set to TRUE. This will produce a set of still images that you can stream together to make a movie.
Default is TRUE. Alternatively (movie=FALSE), you could use this function to get the node labels returned for each permutation starting with a perturbed variable.}
}
\description{
Make a movie of the adaptive walk the diffusion probability method makes in search of a given patient's perturbed variables.
}
\examples{
# Read in any network via its adjacency matrix
tmp = as.matrix(read.table("adjacency_matrix.txt", sep="\\t", header=TRUE))
colnames(tmp) = rownames(tmp)
ig = graph.adjacency(tmp, mode="undirected", weighted=TRUE, add.colnames="name")
V(ig)$name = tolower(V(ig)$name)
adjacency_matrix = list(as.matrix(get.adjacency(ig, attr="weight")))  # Must have this declared as a GLOBAL variable!!!!!
p0=0.1  # 10\% of probability distributed uniformly
p1=0.9  # 90\% of probability diffused based on edge weights in networks
G = vector(mode="list", length=length(V(ig)$name))
names(G) = names(V(ig)$name)
subset.nodes = names(G)[sample(1:length(G), 3)]
mle.getPermMovie(subset.nodes, ig, output_filepath = getwd(), movie=TRUE)
}
\keyword{adaptive}
\keyword{diffusion}
\keyword{event}
\keyword{probability}
\keyword{walk}
