% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.getPermN.r
\name{mle.getPermN}
\alias{mle.getPermN}
\title{Generate the "adaptive walk" node permutations, starting from a given perturbed variable}
\usage{
mle.getPermN(n)
}
\arguments{
\item{n}{- The index (out of a vector of metabolite names) of the permutation you want to calculate.}
}
\description{
This function calculates the node permutation starting from a given perturbed variable in a subset of variables in the background knowledge graph.
}
\examples{
mle.getPermN(n)
}
\keyword{algorithm}
\keyword{diffusion}
\keyword{network}
\keyword{probability}
\keyword{walker}
