% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.getPermN_memory.r
\name{mle.getPermN_memory}
\alias{mle.getPermN_memory}
\title{Generate the "adaptive walk" node permutations, starting from a given perturbed variable}
\usage{
mle.getPermN_memory(S, G)
}
\arguments{
\item{S}{- A character vector of the node names for the subset of nodes you want to encode.}
}
\value{
current_node_set - A character vector of node names in the order they were drawn by the probability diffusion algorithm.
}
\description{
This function calculates the node permutation starting from a given perturbed variable in a subset of variables in the background knowledge graph.
}
\examples{
# Look at main_CTD.r script for full analysis script: https://github.com/BRL-BCM/CTD.
# Get node permutations for graph
perms = list()
for (n in 1:length(G)) {
  print(sprintf("Generating node permutation starting with node \%s", names(G)[n]))
  perms[[n]] = mle.getPermN_memory(n, G)
}
names(perms) = names(G)
}
\keyword{algorithm}
\keyword{diffusion}
\keyword{network}
\keyword{probability}
\keyword{walker}
