% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiNode.getPerms.r
\name{mle.getPerms_memory}
\alias{mle.getPerms_memory}
\title{Generate the "adaptive walk" node permutations, starting from a given perturbed variable}
\usage{
mle.getPerms_memory(S, G, num.misses = NULL)
}
\arguments{
\item{S}{- A character vector of the node names for the subset of nodes you want to encode.}
}
\value{
current_node_set - A character vector of node names in the order they were drawn by the probability diffusion algorithm.
}
\description{
This function calculates the node permutation starting from a given perturbed variable in a subset of variables in the background knowledge graph.
}
\examples{
# Get node permutations for graph
S = names(G)[1:3]
perms = mle.getPermN_memory(S, G)
}
\keyword{algorithm}
\keyword{diffusion}
\keyword{network}
\keyword{probability}
\keyword{walker}
