% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.getPtBSbyK.r
\name{mle.getPtBSbyK}
\alias{mle.getPtBSbyK}
\title{Generate patient-specific bitstrings from adaptive network walk.}
\usage{
mle.getPtBSbyK(data, ptID, perms, kmx)
}
\arguments{
\item{data}{- The matrix that gives the perturbation strength (z-score) for all variables (columns) for each patient (rows).}

\item{ptID}{- The rowname in pvals associated with the patient being processed.}

\item{perms}{- The list of permutations calculated over all possible starting nodes, across all metabolites in data.}

\item{kmx}{- The maximum size of variable sets for which you want to calculate probabilities.}
}
\description{
This function calculates the bitstrings (1 is a hit; 0 is a miss) associated with the adaptive network walk
made by the diffusion algorithm trying to find the variables in the encoded subset, given the background knowledge graph.
}
\examples{
mle.getPtBSbyK(data, ptID, permutationByStartNode, kmax)
}
