% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.kraftMcMillian.r
\name{mle.kraftMcMillian}
\alias{mle.kraftMcMillian}
\title{Apply the Kraft-McMillian Inequality using a specific encoding algorithm.}
\usage{
mle.kraftMcMillian(G, k, memory = FALSE)
}
\arguments{
\item{G}{- A character vector of all node names in the background knowledge graph.}

\item{k}{- The size of the node name subsets of G.}
}
\value{
IA - a list of bitlengths associated with all outcomes in the N choose K outcome space, with the names of the list elements the node names of the encoded nodes
}
\description{
A power analysis of the encoding algorithm using to encode subsets of S in G.
}
