% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.ListMaps_metabolon.r
\name{pathway.ListMaps_metabolon}
\alias{pathway.ListMaps_metabolon}
\title{Get All Metabolites In Metabolon's Pathway Knowledgebase}
\usage{
pathway.ListMaps_metabolon(Pathway.Knowledgebase.Path)
}
\value{
Pathway.Knowledgebase.Path - local path to extdata folder containing metabolon's pathway knowledgebase files.
}
\description{
Get All Metabolites In Metabolon's Pathway Knowledgebase
}
\examples{
metabolon_knowledgebase_path = sprintf("\%s/extdata", find.package("CTD"))
pwys = pathway.ListMaps_metabolon(metabolon_knowledgebase_path)
print(pwys)
}
