% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.chooseKForMDS.r
\name{plot.chooseKForMDS}
\alias{plot.chooseKForMDS}
\title{Metabolite set enrichment analysis (MSEA) using pathway knowledge curated by Metabolon}
\usage{
\method{plot}{chooseKForMDS}(allSimMatrices, diagnoses, diseaseGraph)
}
\arguments{
\item{allSimMatrices}{- A list of all similarity matrices, across all k for a given graph, or across many graphs.}

\item{diagnoses}{- A character vector of diagnostic labels associated with the colnames of each matrix in allSimMatrices.}

\item{diseaseGraph}{- A diagnostic label found in diagnoses that the background knowledge graph was based on.}
}
\description{
A function that returns the pathway enrichment score for all perturbed metabolites in a patient's full metabolomic profile.
}
\examples{
# Look at main_CTD.r script for full analysis script: https://github.com/BRL-BCM/CTD.
# Get patient distances
data_mx.pvals = apply(data_mx, c(1,2), function(i) 2*pnorm(abs(i), lower.tail = FALSE))
res = list()
t = list(ncd=matrix(NA, nrow=ncol(data_mx), ncol=ncol(data_mx)),
         dir=matrix(NA, nrow=ncol(data_mx), ncol=ncol(data_mx)))
rownames(t$ncd) = colnames(data_mx)
colnames(t$ncd) = colnames(data_mx)
rownames(t$dir) = colnames(data_mx)
colnames(t$dir) = colnames(data_mx)
for (i in 1:kmx) {
  res[[i]] = t
}
for (pt in 1:ncol(data_mx)) {
  print(pt)
  ptID = colnames(data_mx)[pt]
  for (pt2 in pt:ncol(data_mx)) {
    ptID2 = colnames(data_mx)[pt2]
    tmp = mle.getPtSim(ptBSbyK[[ptID]], ptID, ptBSbyK[[ptID2]], ptID2, data_mx, perms)
    for (k in 1:kmx) {
      res[[k]]$ncd[ptID, ptID2] = tmp$NCD[k]
      res[[k]]$dir[ptID, ptID2] = tmp$dirSim[k]
      res[[k]]$ncd[ptID2, ptID] = tmp$NCD[k]
      res[[k]]$dir[ptID2, ptID] = tmp$dirSim[k]
    }
  }
}
res_tmp = lapply(res, function(i) i$ncd)
}
