% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.getPathwayIgraph.r
\name{plot.getPathwayIgraph}
\alias{plot.getPathwayIgraph}
\title{plot.getPathwayIgraph}
\usage{
plot.getPathwayIgraph(input, Pathway.Name)
}
\arguments{
\item{input}{- A list object of parameters (esp. from R shiny app). Required parameters are ptIDs, diagClass and pathwayMapId.}

\item{Pathway.Name}{- The name of the pathway map for which you want the topological information.}
}
\value{
template.ig - Igraph object of selected pathway map.
}
\description{
plot.getPathwayIgraph
}
\examples{
data(Miller2015)
# Input is supplied by R shiny app, but you can hard code parameters as a list object, too, to test functionality.
input = list()
input$ptIDs = colnames(Miller2015)[4]
input$diagClass = "paa"
input$pathwayMapId = "All"
ig = plot.getPathwayIgraph(input, Miller2015)
# Returns a blank template for selected pathway.
plot.igraph(ig, edge.arrow.size = 0.01)
}
