% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hmSim.r
\name{plot.hmSim}
\alias{plot.hmSim}
\title{Generate heatmap plot of patient similarity matrix.}
\usage{
plot.hmSim(patientSim, path=getwd(), diagnoses=NULL)
}
\arguments{
\item{patientSim}{- The patient similarity matrix.}

\item{path}{- The filepath to a directory in which you want to store the .png file.}

\item{diagnoses}{- A character vector of diagnostic labels associated with the rownames of patientSim.}
}
\description{
This function plots a heatmap of a patient similarity matrix.
}
\examples{
plot.hmSim(patientSim, path=getwd(), diagnoses)
}
