% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pathwayMap_CaseCntl.r
\name{plot.pathwayMap_CaseCntl}
\alias{plot.pathwayMap_CaseCntl}
\title{Generate pathway map with patient perturbation data superimposed.}
\usage{
plot.pathwayMap_CaseCntl(Pathway, groupName, pvalues, out.path, SVG=TRUE)
}
\arguments{
\item{Pathway}{- The name of the pathway map you want to plot patient data on.}

\item{groupName}{- An identifier string associated with the group being compared to a "baseline" condition.}

\item{pvalues}{- A named vector of metabolites with corresponding p-values (e.g., from a t-test).}

\item{out.path}{- The directory in which you want to store image files.}

\item{SVG}{- Save as SVG or PNG? If SVG is TRUE, then an SVG image is saved. If FALSE, a PNG is saved.}
}
\description{
Generate pathway map with patient perturbation data superimposed.
}
\examples{
pwys = pathway.ListMaps_metabolon()
data(Miller2015)
Miller2015 = Miller2015[,grep("IEM", colnames(Miller2015))]
type = diagnoses$diagnosis
sdd = apply(Miller2015[,which(type=="Argininemia")], 1, sd)
Miller2015 = Miller2015[-which(sdd==0),]
pvals = unlist(apply(Miller2015, 1, function(i) t.test(i[which(type=="Argininemia")], i[which(type=="No biochemical genetic diagnosis")])$p.value))
names(pvals) = rownames(Miller2015)
plot.pathwayMap_CaseCntl("allPathways", groupName="Argininemia", pvalues=pvals, out.path=getwd(), SVG=TRUE)
}
