% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.getMSEA_Metabolon.R
\name{shiny.getMSEA_Metabolon}
\alias{shiny.getMSEA_Metabolon}
\title{Metabolite set enrichment analysis (MSEA) using pathway knowledge curated by Metabolon}
\usage{
shiny.getMSEA_Metabolon(input)
}
\description{
A function that returns the pathway enrichment score for all perturbed metabolites in a patient's full metabolomic profile.
}
\examples{
pathway.data = shiny.getMSEA_Metabolon(input)
}
