% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.getORA_Metabolon.R
\name{shiny.getORA_Metabolon}
\alias{shiny.getORA_Metabolon}
\title{Metabolite set enrichment analysis (MSEA) (using a hypergeometric test) using pathway knowledge curated by Metabolon}
\usage{
shiny.getORA_Metabolon(input)
}
\description{
A function that returns the pathway enrichment score for all perturbed metabolites in a patient's full metabolomic profile.
}
\examples{
pathway.data = shiny.getORA_Metabolon(met.profile, threhold=3, "z-score", NULL)
}
