% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.entropyFunction.r
\name{stat.entropyFunction}
\alias{stat.entropyFunction}
\title{Entropy of a bit-string}
\usage{
stat.entropyFunction(bitString)
}
\arguments{
\item{bitString}{- A vector of 0's and 1's.}
}
\value{
e - a floating point percentage, between 0 and 1.
}
\description{
The entropy of a bitstring (ex: 1010111000) is calculated.
}
\examples{
stat.entropyFunction(c(1,0,0,0,1,0,0,0,0,0,0,0,0))       # Output: 0.6193822
stat.entropyFunction(c(1,1,1,1,1,1,1,1,0,0,0,0,0,0,0,0)) # Output: 1
stat.entropyFunction(c(1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1)) # Output: 0
}
