% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.fishersMethod.r
\name{stat.fishersMethod}
\alias{stat.fishersMethod}
\title{Fisher's Combined P-value}
\usage{
stat.fishersMethod(x)
}
\arguments{
\item{x}{- A vector of p-values (floating point numbers).}
}
\value{
a floating point number, a combined p-value using Fisher's 
method.
}
\description{
Fisher's combined p-value, used to combine the results of individual 
statistical tests into an overall hypothesis.
}
\examples{
stat.fishersMethod(c(0.2,0.1,0.3))   # Output: 0.1152162
}
