% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.entropyFunction.r
\name{stats.entropyFunction}
\alias{stats.entropyFunction}
\title{Entropy of a bit-string}
\usage{
stats.entropyFunction(bitString)
}
\arguments{
\item{x}{- A vector of 0's and 1's.}
}
\description{
The entropy of a bitstring (ex: 1010111000) is calculated.
}
\examples{
stats.entropyFunction(c(1,0,0,0,1,0,0,0,0,0,0,0,0))
> 0.6193822
stats.entropyFunction(c(1,1,1,1,1,1,1,1,0,0,0,0,0,0,0,0))
> 1
stats.entropyFunction(c(1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1))
> 0
}
