% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.fishersMethod.r
\name{stats.fishersMethod}
\alias{stats.fishersMethod}
\title{Fisher's Combined P-value}
\usage{
stats.fishersMethod(x)
}
\arguments{
\item{x}{- A vector of floating point numbers.}
}
\description{
Fisher's combined p-value, used to combine the results of individual statistical tests into an overall hypothesis.
}
\examples{
stats.fishersMethod(c(0.2,0.1,0.3))
> 0.1152162
}
