% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.getORA_Metabolon.r
\name{stats.getORA_Metabolon}
\alias{stats.getORA_Metabolon}
\title{Metabolite set enrichment analysis (MSEA) (using a hypergeometric test) using pathway knowledge curated by Metabolon}
\usage{
stats.getORA_Metabolon(met.profile, threshold = 3, type = "zscore",
  gene.profile = NULL)
}
\arguments{
\item{met.profile}{- A character vector of a patient's metabolomic profile, including KEGG IDs
and the associated z-score or p-value describing the level of the metabolite compared to controls.}

\item{threshold}{- A cutoff to select metabolites with a zscore > threshold or < -1*threshold.}

\item{type}{- Either "p-value" or "z-score".}

\item{gene.profile}{- Default set to NULL, meaning the default enrichment analysis only considers metabolites. However, if you have gene data, too, set
this parameter to a character vector of the gene names with found variants in the patient's record. Gene IDs must be converted to Entrez Identifiers.}
}
\description{
A function that returns the pathway enrichment score for all perturbed metabolites in a patient's full metabolomic profile.
}
\examples{
pathway.data = stats.getORA_Metabolon(met.profile, threhold=3, "z-score", NULL)
}
