% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.iteratedLog2.r
\name{stats.iteratedLog2}
\alias{stats.iteratedLog2}
\title{Iterated Logarithm (Base 2)}
\usage{
stats.iteratedLog2(num)
}
\arguments{
\item{num}{- An integer.}
}
\description{
This function calculates the number of times the logarithm function (base 2) must be iteratively applied before the result is less than or equal to 1.
}
\examples{
iteratedLogarithm(4)
2
}
