% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.Thistlethwaite2020.r
\docType{data}
\name{Thistlethwaite2020}
\alias{Thistlethwaite2020}
\title{Clinical Diagnosis of Metabolic Disorders using Untargeted Metabolomic Profiling and Disease-specific Networks Learned from Patient Data}
\source{
\href{https://genboree.org/Metabolomics-Data-Portal/}{Dataset}
}
\usage{
data(Thistlethwaite2020)
}
\value{
cohorts_coded - A list object where elements contain all patient IDs associated with a given diagnosis.

data_zscore - The data matrix (metabolite features are rows, patient observations are columns) for 545 untargeted metabolomics 
                      patient samples, alongside metabolite annotations.
}
\description{
A meta-analysis of previous untargeted metabolomics studies describing 16 
unique inborn errors of metabolism.
}
\examples{
require(CTD)
data(Thistlethwaite2020)
}
\references{
L.R. Thistlethwaite, et al. 2020. In review.
}
\keyword{datasets}
