% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.Wangler2017.r
\docType{data}
\name{Wangler2017}
\alias{Wangler2017}
\title{A metabolomic map of Zellweger spectrum disorders reveals novel disease biomarkers.}
\format{
Wangler2017 - The data matrix (metabolite features are rows, patient observations are columns) for 19 untargeted metabolomics 
                      patient samples, alongside metabolite annotations.
}
\source{
\href{https://static-content.springer.com/esm/art:10.1038/gim.2017.262/MediaObjects/41436_2018_207_MOESM2_ESM.xlsx}{Dataset}
}
\usage{
data(Wangler2017)
}
\description{
Global metabolic profiling obtained by untargeted mass spectrometry-based
metabolomic platform for the detection of novel and known inborn errors of
metabolism.  This untargeted approach collected z-score values for >650 unique
compounds from human plasma.  Data set contains 19 individual plasma samples with
confirmed biallelic pathogenic variants in the PEX1 gene. These samples revealed
elevations in pipecolic acid and long-chain lysophosphatidylcholines, as well as
an unanticipated reduction in multiple sphingomyelin species.
}
\examples{
data(Wangler2017)
}
\references{
M.F. Wangler, et al. Genetics in Medicine, 2018, 00
(\href{https://www.ncbi.nlm.nih.gov/pubmed/29419819}{PubMed})
}
\keyword{datasets}
