% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctdTag.R
\name{ctdTag}
\alias{ctdTag}
\title{Tag CTD features}
\usage{
ctdTag(file, dbname = NULL, taglist = NULL)
}
\arguments{
\item{file}{name of file containing CTD data, which must be in a format
handled by \code{\link[oce:read.oce]{oce::read.oce()}}.}

\item{dbname}{optional character value specifying the name of a sqlite
database used to hold tagging information.  If not provided, a file name
is constructed with \code{\link[=getDatabaseName]{getDatabaseName()}}.
If this file does not exist, it is created, and a table named
\code{taglist} is constructed in it, according to the \code{taglist} parameter.
However, if the file already exists, then its \code{taglist} table is
not altered.}

\item{taglist}{optional \link{list} containing (FILL IN).}
}
\value{
either an empty string, if the procedure worked, or a description
of the problem, otherwise.
}
\description{
Tag CTD features
}
\author{
Dan Kelley
}
