% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctdTag.R
\name{ctdTag}
\alias{ctdTag}
\title{Tag CTD features}
\usage{
ctdTag(
  file,
  dbname = NULL,
  taglist = NULL,
  height = 400,
  clickDistanceCriterion = 0.02,
  debug = 0
)
}
\arguments{
\item{file}{character value specifying the name of a file containing CTD data,
which must be in a format handled by \code{\link[oce:read.oce]{oce::read.oce()}}.}

\item{dbname}{optional character value specifying the name of a sqlite
database used to hold tagging information.  If not provided, a file name
is constructed with \code{\link[=getDatabaseName]{getDatabaseName()}}.
If this file does not exist, it is created, and a table named
\code{taglist} is constructed in it, according to the \code{taglist} parameter.
However, if the file already exists, then its \code{taglist} table is
not altered.}

\item{taglist}{optional \link{list} containing (FILL IN).}

\item{height}{height of the plot in pixels.}

\item{clickDistanceCriterion}{numeric value controlling whether a mouse click
selects, or deselects a point.  If the ratio of the distance from the mouse
location to the nearest point exceeds \code{clickDistanceCriterion} times
the span of the current view, then that nearest point is selected. Otherwise,
any existing selected point is deselected.}

\item{debug}{integer indicating whether to print some informative information
to the R console.  Use 0 for mainly silent processing, 1 for more information, etc.}
}
\value{
\code{\link[=ctdTag]{ctdTag()}} returns either an empty string, if the procedure worked, or a description
of the problem, otherwise.
}
\description{
Tag CTD features
}
\examples{
\dontrun{
library(ctd)
file <- "~/data/arctic/beaufort/2012/d201211_0047.cnv"
ctdTag(file=file, height=550)
}

}
\author{
Dan Kelley
}
