% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.naivePruning.r
\name{graph.naivePruning}
\alias{graph.naivePruning}
\title{Pruning edges from disease differential network that also occur in reference-only network.}
\usage{
graph.naivePruning(ig_dis, ig_ref)
}
\arguments{
\item{ig_dis}{- The igraph object associated with the disease+reference trained differential interaction network.}

\item{ig_ref}{- The igraph object associated with the reference-only trained interaction network.}
}
\value{
ig_pruned - The pruned igraph object of the disease+reference differential interaction network, with reference edges subtracted.
}
\description{
Recursively diffuse probability from a starting node based on the connectivity of the background knowledge graph, representing the likelihood that a variable will be
        most influenced by a perturbation in the starting node.
}
\examples{
ig_pruned=graph.naivePruning(ig_dis, ig_ref)
}
