% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.getEncodingLength.r
\name{mle.getEncodingLength}
\alias{mle.getEncodingLength}
\title{Minimum encoding length (MLE)}
\usage{
mle.getEncodingLength(bs, pvals, ptID)
}
\arguments{
\item{bs}{- A list of bitstrings associated with a given patient's perturbed variables.}

\item{pvals}{- The matrix that gives the perturbation strength significance for all variables (columns) for each patient (rows)}

\item{ptID}{- The row name in data.pvals corresponding to the patient you specifically want encoding information for.}
}
\description{
This function calculates the mininmum encoding length associated with a subset of variables given a background knowledge graph.
}
\examples{
mle.getEncodingLength(optBS, data.pvals, ptID)
}
\keyword{encoding}
\keyword{length}
\keyword{minimum}
