% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.getEncodingLength.r
\name{mle.getEncodingLength}
\alias{mle.getEncodingLength}
\title{Minimum encoding length (MLE)}
\usage{
mle.getEncodingLength(bs, pvals, ptID, G)
}
\arguments{
\item{bs}{- A list of bitstrings associated with a given patient's perturbed variables.}

\item{pvals}{- The matrix that gives the perturbation strength significance for all variables (columns) for each patient (rows)}

\item{ptID}{- The row name in data.pvals corresponding to the patient you specifically want encoding information for.}

\item{G}{- A list of probabilities with list names being the node names of the background graph.}
}
\value{
df - a data.frame object, for every bitstring provided in bs input parameter, a row is returned with the following data:
             the patientID; the bitstring evaluated where T denotes a hit and 0 denotes a miss; the subsetSize, or the number of
             hits in the bitstring; the individual p-values associated with the variable's perturbations, delimited by '/';
             the combined p-value of all variables in the set using Fisher's method; Shannon's entropy, IS.null;
             the minimum encoding length IS.alt; and IS.null-IS.alt, the d.score.
}
\description{
This function calculates the mininmum encoding length associated with a subset of variables given a background knowledge graph.
}
\examples{
# Identify the most significant subset per patient, given the background graph
data_mx.pvals = t(apply(data_mx, c(1,2), function(i) 2*pnorm(abs(i), lower.tail = FALSE)))
for (pt in 1:ncol(data_mx)) {
    ptID = colnames(data_mx)[pt]
    res = mle.getEncodingLength(ptBSbyK[[ptID]], data_mx.pvals, ptID, G)
    res = res[order(res[,"d.score"], decreasing=TRUE),]
    print(res)
}
}
\keyword{encoding}
\keyword{length}
\keyword{minimum}
