% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.getPatientSimilarity.r
\name{mle.getPatientSimilarity}
\alias{mle.getPatientSimilarity}
\title{Patient similarity using mutual information MLE metric of patients' most modular, perturbed subsets.}
\usage{
mle.getPatientSimilarity(p1.optBS, ptID, p2.optBS, ptID2, data, data.pvals)
}
\arguments{
\item{p1.optBS}{- The optimal bitstring associated with patient 1.}

\item{ptID}{- The identifier associated with patient 1's sample.}

\item{p2.optBS}{- The optimal bitstring associated with patient 2.}

\item{data}{- The matrix that gives the perturbation strength (z-scores) for all variables (columns) for each patient (rows).}

\item{ptID}{- The identifier associated with patient 2's sample.}

\item{pvals}{- The matrix that gives the perturbation strength significance (p-value) for all variables (columns) for each patient (rows).}
}
\description{
This function calculates the universal distance between patients, using a mutual information metric, where self-information comes from the minimum encoding length of each patient's encoded modular perturbations in the background knowledge graph.
}
