% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.getPermMovie.r
\name{mle.getPermMovie}
\alias{mle.getPermMovie}
\title{Capture the movement of the adaptive walk of the diffusion probability method.}
\usage{
mle.getPermMovie(patient.sig.nodes, patient, ig, output_filepath,
  movie = TRUE)
}
\arguments{
\item{patient.sig.nodes}{- The subset of variables, S, in a background graph, G. These are the perturbed molecular phenotype variables in a given patient's profile.}

\item{patient}{- The column number in data.pvals associated with the patient being examined.}

\item{ig}{- The igraph object associated with the background knowledge graph.}

\item{movie}{- If you want to make a movie, set to TRUE. This will produce a set of still images that you can stream together to make a movie.
Default is TRUE. Alternatively (movie=FALSE), you could use this function to get the node labels returned for each permutation starting with a perturbed variable.}
}
\description{
Make a movie of the adaptive walk the diffusion probability method makes in search of a given patient's perturbed variables.
}
\examples{
mle.getPermMovie(patient.sig.nodes, patient, ig)
}
\keyword{adaptive}
\keyword{diffusion}
\keyword{event}
\keyword{probability}
\keyword{walk}
