% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiNode.getPtBSbyK.r
\name{multiNode.getPtBSbyK}
\alias{multiNode.getPtBSbyK}
\title{Generate patient-specific bitstrings from adaptive network walk.}
\usage{
multiNode.getPtBSbyK(S, ranks)
}
\arguments{
\item{S}{- A character vector of node names describing the node subset to be encoded.}

\item{ranks}{- The list of node ranks calculated over all possible nodes in S, starting with each node in subset of interest.}
}
\value{
pt.byK - a list of bitstrings, with the names of the list elements the node names of the encoded nodes
}
\description{
This function calculates the bitstrings (1 is a hit; 0 is a miss) associated with the adaptive network walk
made by the diffusion algorithm trying to find the variables in the encoded subset, given the background knowledge graph.
}
\examples{
# Get bitstrings associated with each patient's top kmx variable subsets
kmx = 15
ptBSbyK = list()
for (pt in 1:ncol(data_mx)) {
  S = data_mx[order(abs(data_mx[,pt]), decreasing=TRUE),pt][1:kmx]
  ptBSbyK[[ptID]] = multiNode.getPtBSbyK(S, ranks)
}
}
