% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.ListMaps_metabolon.r
\name{pathway.ListMaps_metabolon}
\alias{pathway.ListMaps_metabolon}
\title{Get All Metabolites In Metabolon's Pathway Knowledgebase}
\usage{
pathway.ListMaps_metabolon()
}
\value{
pwys - List of pathway maps curated by Metabolon's Metabolync.
}
\description{
Get All Metabolites In Metabolon's Pathway Knowledgebase
}
\examples{
pwys = pathway.ListMaps_metabolon()
print(pwys)
}
