% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hmSim.r
\name{plot.hmSim}
\alias{plot.hmSim}
\title{Generate heatmap plot of patient similarity matrix.}
\usage{
\method{plot}{hmSim}(simMat, path, diagnoses = NULL)
}
\arguments{
\item{simMat}{- The patient similarity matrix.}

\item{path}{- The filepath to a directory in which you want to store the .png file.}

\item{diagnoses}{- A character vector of diagnostic labels associated with the rownames of simMat.}
}
\description{
This function plots a heatmap of a patient similarity matrix.
}
\examples{
# Look at main_CTD.r script for full analysis script: https://github.com/BRL-BCM/CTD.
plot.hmSim(patientSim, path=getwd(), diagnoses)
}
