% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.knnSim.r
\name{plot.knnSim}
\alias{plot.knnSim}
\title{Visualize the confusion matrix using nearest neighbor as classification model.}
\usage{
plot.knnSim(patientSim, diagnoses, diag)
}
\arguments{
\item{patientSim}{- The patient similarity matrix.}

\item{diagnoses}{- A character vector of diagnostic labels associated with the rownames of patientSim.
The names of this vector are patient IDs, and values are diagnostic labels.}

\item{diag}{- The diagnosis associated with positive controls in your data.}
}
\value{
p - a plotly scatter plot colored by provided diagnostic labels.
}
\description{
Visualize the confusion matrix using nearest neighbor as classification model.
}
\examples{
# if you have diagnostic labels associated with the colnames(data_mx), send them using diagnoses parameter
p = plot.knnSim(patientSim, diagnoses, diag="diseased")
p
}
