% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mdsSim.r
\name{plot.mdsSim}
\alias{plot.mdsSim}
\title{View patient clusters using multi-dimensional scaling.}
\usage{
\method{plot}{mdsSim}(patientSim, diagnoses, k, diag)
}
\arguments{
\item{diagnoses}{- A character vector of diagnostic labels associated with the rownames of simMat.}

\item{k}{- The number of dimension you want to plot your data using multi-dimensional scaling.}

\item{diag}{- The diagnosis associated with positive controls in your data.}

\item{simMat}{- The patient similarity matrix.}
}
\value{
p - a plotly scatter plot colored by provided diagnostic labels.
}
\description{
This function plots the provided patient similarity matrix in a lower dimensional space
using multi-dimensional scaling, which is well suited for similarity metrics.
}
\examples{
# Look at main_CTD.r script for full analysis script: https://github.com/BRL-BCM/CTD.
# if you have diagnostic labels associated with the colnames(data_mx), send them using diagnoses parameter
p = plot.mdsSim(patientSim, diagnoses, k=2, diag="diseased")
p
p = plot.mdsSim(patientSim, diagnoses, k=3, diag="diseased")
p
}
