% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mdsSim.r
\name{plot.mdsSim}
\alias{plot.mdsSim}
\title{View patient clusters using multi-dimensional scaling.}
\usage{
\method{plot}{mdsSim}(simMat, diagnoses, k, diag)
}
\arguments{
\item{simMat}{- The patient similarity matrix.}

\item{diagnoses}{- A character vector of diagnostic labels associated with the rownames of simMat.}

\item{k}{- The number of dimension you want to plot your data using multi-dimensional scaling.}

\item{diag}{- The diagnosis associated with positive controls in your data.}
}
\description{
This function plots the provided patient similarity matrix in a lower dimensional space
using multi-dimensional scaling, which is well suited for similarity metrics.
}
\examples{
plot.mdsSim(simMat, path)
}
