% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pathwayMap.r
\name{plot.pathwayMap}
\alias{plot.pathwayMap}
\title{Generate pathway map with patient perturbation data superimposed.}
\usage{
\method{plot}{pathwayMap}(Pathway.Name, PatientID, patient.zscore,
  zscore.threshold, scalingFactor, outputFilePath, SVG = TRUE)
}
\arguments{
\item{Pathway.Name}{- The name of the pathway map you want to plot patient data on.}

\item{PatientID}{- An identifier string associated with the patient.}

\item{patient.zscore}{- A named vector of metabolites with corresponding z-scores.}

\item{zscore.threshold}{- Plot all z-scores > or < this threshold.}

\item{scalingFactor}{- Integer associated with increase in node size.}

\item{outputFilePath}{- The directory in which you want to store image files.}

\item{SVG}{- Save as SVG or PNG? If SVG is TRUE, then an SVG image is saved. If FALSE, a PNG is saved.}
}
\description{
Generate pathway map with patient perturbation data superimposed.
}
\examples{
plot.pathwayMap(simMat, path)
}
