% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pathwayMap.r
\name{plot.pathwayMap}
\alias{plot.pathwayMap}
\title{Generate pathway map with patient perturbation data superimposed.}
\usage{
plot.pathwayMap(Pathway, ptID, pt.zscore, zscore.threshold, scale, out.path, SVG=TRUE)
}
\arguments{
\item{Pathway}{- The name of the pathway map you want to plot patient data on.}

\item{ptID}{- An identifier string associated with the patient.}

\item{pt.zscore}{- A named vector of metabolites with corresponding z-scores.}

\item{zscore.threshold}{- Plot all z-scores > or < this threshold.}

\item{scale}{- Integer associated with increase in node size.}

\item{out.path}{- The directory in which you want to store image files.}

\item{SVG}{- Save as SVG or PNG? If SVG is TRUE, then an SVG image is saved. If FALSE, a PNG is saved.}
}
\description{
Generate pathway map with patient perturbation data superimposed.
}
\examples{
Pathway = pathway.ListMaps_metabolon()
data(Miller2015)
Miller2015 = Miller2015[,grep("IEM", colnames(Miller2015))]
ptID = colnames(Miller2015)[1]
pt.zscore = Miller2015[,1]
plot.pathwayMap(Pathway[1], ptID, pt.zscore, zscore.threshold, scale=1, out.path=getwd(), SVG=TRUE)
}
