% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleNode.getPtBSbyK.r
\name{singleNode.getPtBSbyK}
\alias{singleNode.getPtBSbyK}
\title{Generate patient-specific bitstrings from the fixed, single-node network walker.}
\usage{
singleNode.getPtBSbyK(S, ranks, num.misses = NULL)
}
\arguments{
\item{S}{- A character vector of node names describing the node subset to be encoded.}

\item{ranks}{- The list of node ranks calculated over all possible nodes, starting with each node in subset of interest.}

\item{num.misses}{- The number of misses tolerated by the network walker before path truncation occurs.}
}
\value{
pt.byK - a list of bitstrings, with the names of the list elements the node names of the encoded nodes
}
\description{
This function calculates the bitstrings (1 is a hit; 0 is a miss) associated with the fixed, single-node network walker
trying to find the variables in the encoded subset, given the background knowledge graph.
}
\examples{
# Load in your profiling data (rows are compounds, columns are samples)
data_mx = read.table("your_profiling_data.txt", sep="\\t", header=TRUE)
# Get bitstrings associated with each patient's top kmx variable subsets
kmx = 15
ptBSbyK = list()
for (pt in 1:ncol(data_mx)) {
  S = data_mx[order(abs(data_mx[,pt]), decreasing=TRUE),pt][1:kmx]
  ptBSbyK[[ptID]] = singleNode.getPtBSbyK(S, ranks)
}
}
