% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.Thistlethwaite2020.r
\docType{data}
\name{Thistlethwaite2020}
\alias{Thistlethwaite2020}
\title{Thistlethwaite et al. (2020)}
\format{
Thistlethwaite2020 - A data frame with 1364 metabolite 
features as rows and 545 untargeted metabolomics patient samples
as columns, alongside 8 metabolite annotations.
}
\source{
\href{https://genboree.org/Metabolomics-Data-Portal/}{Dataset}
}
\usage{
data(Thistlethwaite2020)
}
\description{
Clinical Diagnosis of Metabolic Disorders using Untargeted 
Metabolomic Profiling and Disease-specific Networks Learned from 
Patient Data. A meta-analysis of previous untargeted metabolomics
studies describing 16 unique inborn errors of metabolism.
}
\examples{
data(Thistlethwaite2020)
}
\references{
L.R. Thistlethwaite, et al. 2020. In review.
}
\keyword{datasets}
