% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.Wangler2017.r
\docType{data}
\name{Wangler2017}
\alias{Wangler2017}
\title{Wangler et al. (2017)}
\format{
Wangler2017 - The data matrix (metabolite features are
rows, patient observations are columns) for 19 untargeted metabolomics 
patient samples, alongside metabolite annotations.
}
\source{
Electronic supplementary material, Supplementary Tables
(\href{https://tinyurl.com/y4zus9l2}{Dataset})
}
\usage{
data(Wangler2017)
}
\description{
A metabolomic map of Zellweger spectrum disorders reveals 
novel disease biomarkers. Global metabolic profiling obtained by 
untargetedmass spectrometry-based metabolomic platform for the 
detectionof novel and known inborn errors of metabolism.  This 
untargeted approach collected z-score values for >650 unique compounds 
fromhuman plasma.  Data set contains 19 individual plasma samples with
confirmed biallelic pathogenic variants in the PEX1 gene. These 
samples revealed elevations in pipecolic acid and long-chain
lysophosphatidylcholines, as well as an unanticipated reduction
in multiple sphingomyelin species.
}
\examples{
data(Wangler2017)
}
\references{
M.F. Wangler, et al. (2018). A metabolomic map of Zellweger 
spectrum disorders reveals novel disease biomarkers. Genetics in 
Medicine, 20: 1274-1283.
(\href{https://doi.org/10.1038/gim.2017.262}{Article})
}
\keyword{datasets}
